Support arrays-as-dict, progress visibility schema
The Jan-Willem school of writing quantities is to use the dictionary
representation of data arrays... So where we expect an array but find a
dictionary, the right thing to do is to use xarray.DataArray.from_dict in order
to get an array that we can check. The entire thing is *slightly* awkward
because coordinates might now change depending on context, we might need a
better solution for that.

This also makes the numpy dtype check a bit more lenient: Asking for
"U" (strings) should allow strings of fixed size (i.e. "U4").