Add lofar mode
Adds read_col_conversion_dask which creates a lazy representation of DataArray. Each lazy chunk contains a column read and a reshaping step. The input argument lofar_read_size is used to control both the size of contiguous reads from the measurement set and the size of chunks on the zarr store